팀 나가기 로직을 useTeamManagement 훅으로 분리
